+++ packages/market-service/src/coingecko/coingecko.ts
@@ -38,20 +38,34 @@ type CoinGeckoAssetData = {
 }
 
 export class CoinGeckoMarketService implements MarketService {
-  baseUrl = 'https://api.coingecko.com/api/v3'
+  baseUrl: string
+  APIKey: string
+
+  constructor(args: { coinGeckoAPIKey: string }) {
+    const { coinGeckoAPIKey } = args
+    this.APIKey = coinGeckoAPIKey
+    // if we have a key - use the pro- api
+    this.baseUrl = `https://${this.APIKey ? 'pro-' : ''}api.coingecko.com/api/v3`
+  }
 
   private readonly defaultGetByMarketCapArgs: FindAllMarketArgs = {
     count: 2500
   }
 
-  findAll = async (args?: FindAllMarketArgs) => {
+  private maybeAddAPIKey = (): string => {
+    return this.APIKey ? `&x_cg_pro_api_key=${this.APIKey}` : ''
+  }
+
+  async findAll(args?: FindAllMarketArgs) {
     const argsToUse = { ...this.defaultGetByMarketCapArgs, ...args }
     const { count } = argsToUse
     const perPage = count > 250 ? 250 : count
     const pages = Math.ceil(bnOrZero(count).div(perPage).toNumber())
 
     const urlAtPage = (page: number) =>
-      `${this.baseUrl}/coins/markets?vs_currency=usd&order=market_cap_desc&per_page=${perPage}&page=${page}&sparkline=false`
+      `${
+        this.baseUrl
+      }/coins/markets?vs_currency=usd&order=market_cap_desc&per_page=${perPage}&page=${page}&sparkline=false${this.maybeAddAPIKey()}`
     const pageCount = Array(pages)
       .fill(0)
       .map((_v, i) => i + 1)
@@ -91,7 +105,7 @@ export class CoinGeckoMarketService implements MarketService {
     }
   }
 
-  findByAssetId = async ({ assetId }: MarketDataArgs): Promise<MarketData | null> => {
+  async findByAssetId({ assetId }: MarketDataArgs): Promise<MarketData | null> {
     try {
       if (!adapters.assetIdToCoingecko(assetId)) return null
       const { chainNamespace, assetReference } = fromAssetId(assetId)
@@ -100,7 +114,7 @@ export class CoinGeckoMarketService implements MarketService {
       const contractUrl = isToken ? `/contract/${assetReference}` : ''
 
       const { data }: { data: CoinGeckoAssetData } = await axios.get(
-        `${this.baseUrl}/coins/${id}${contractUrl}`
+        `${this.baseUrl}/coins/${id}${contractUrl}?${this.maybeAddAPIKey()}`
       )
 
       const currency = 'usd'
@@ -125,10 +139,10 @@ export class CoinGeckoMarketService implements MarketService {
     }
   }
 
-  findPriceHistoryByAssetId = async ({
+  async findPriceHistoryByAssetId({
     assetId,
     timeframe
-  }: PriceHistoryArgs): Promise<HistoryData[]> => {
+  }: PriceHistoryArgs): Promise<HistoryData[]> {
     if (!adapters.assetIdToCoingecko(assetId)) return []
     try {
       const { chainNamespace, assetReference } = fromAssetId(assetId)
@@ -170,7 +184,7 @@ export class CoinGeckoMarketService implements MarketService {
 
       const currency = 'usd'
       const { data: historyData } = await axios.get<CoinGeckoHistoryData>(
-        `${url}/market_chart/range?id=${id}&vs_currency=${currency}&from=${from}&to=${to}`
+        `${url}/market_chart/range?id=${id}&vs_currency=${currency}&from=${from}&to=${to}${this.maybeAddAPIKey()}`
       )
       return historyData?.prices?.reduce<HistoryData[]>((acc, data) => {
         const date = data[0]