+++ packages/swapper/src/swappers/thorchain/types.ts
@@ -1,3 +1,5 @@
+import { ChainAdapterManager } from '@shapeshiftoss/chain-adapters'
+
 export type MidgardResponse = {
   asset: string
   assetDepth: string
@@ -11,4 +13,18 @@ export type MidgardResponse = {
   synthUnits: string
   units: string
   volume24h: string
+}
+
+export type InboundResponse = {
+  chain: string
+  pub_key: string
+  address: string
+  halted: boolean
+  gas_rate: string
+  router?: string
+}
+
+export type ThorchainSwapperDeps = {
+  midgardUrl: string
+  adapterManager: ChainAdapterManager
 }